requirements.txt
spotipy>=2.22.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.25.0
requests>=2.31.0
plotly>=5.15.0
python-dotenv>=1.0.0
kivy>=2.2.0
kivymd>=1.1.1
pillow>=7.1.0,<10
